rename folder 'util' -> 'sequence' + change error message for infinite sequence
