+++ src/test/java/net/datafaker/sequence/Data.java
@@ -1,4 +1,4 @@
-package net.datafaker.util;
+package net.datafaker.sequence;
 
 import net.datafaker.providers.base.BaseFaker;
